封装方法、

+++ app/Libs/QiniuCloud.php
@@ -30,21 +30,23 @@ class QiniuCloud
 	 * TODO 上传七牛云
 	 * 
 	 * @param  string $filepath [本地路径]
-	 * @param  [type] $filename [文件名]
+	 * @param  [type] $key      [文件名]
 	 * @return [type]           [description]
 	 */
-	public function qiniu_upload($filepath='',$filename='')
+	public function qiniu_upload($filePath='',$key='default.jpg')
 	{
         try{
 	        // 要上传文件的本地路径
-	        $filePath = 'D:\phpStudy\WWW\spider\public\image\defaul1123t.jpg';
+	        $filePath = $filePath;
+	        if(!$filePath){
+	        	return 'File Path not be null';
+	        }
 	        // 上传到七牛后保存的文件名
-	        $key = 'bk2.png';
+	        $key = $filename;
 	        // 初始化 UploadManager 对象并进行文件的上传。
 	        $uploadMgr = new UploadManager();
 	        // 调用 UploadManager 的 putFile 方法进行文件的上传。
 	        list($ret, $err) = $uploadMgr->putFile($this->token, $key, $filePath);
-	        echo "\n====> putFile result: \n";
 	        unlink($filePath);
 	        if ($err !== null) {
 	            return($err);
